update to use the mask
