+++ ui/src/style/tooltip.less
@@ -64,12 +64,8 @@
 
 .tooltip-sentence__container {
   position: fixed;
-  top: 92vh;
-  right: 3vw;
-
-  .tooltip-sentence__card {
-    width: 300px;
-  }
+  top: 85vh;
+  right: 2vw;
 }
 
 /**
